+++ packages/yourloops/pages/signup/signup-stepper.tsx
@@ -26,22 +26,19 @@
  */
 
 import React, { type FunctionComponent, useState } from 'react'
-import { useNavigate } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-
 import { type Theme } from '@mui/material/styles'
 import { makeStyles } from 'tss-react/mui'
 import Box from '@mui/material/Box'
-import Button from '@mui/material/Button'
 import Stepper from '@mui/material/Stepper'
 import Step from '@mui/material/Step'
 import StepLabel from '@mui/material/StepLabel'
 import Typography from '@mui/material/Typography'
-
 import SignUpProfileForm from './signup-profile-form'
 import SignUpConsent from './signup-consent'
 import { useAuth } from '../../lib/auth'
 import SignUpAccountSelector from './signup-account-selector'
+import { SignupCompleteMessage } from './signup-complete-message'
 
 export interface SignUpFormProps {
   handleBack: () => void
@@ -57,17 +54,12 @@ const useStyles = makeStyles()((theme: Theme) => ({
   }
 }))
 
-const SignUpStepper: FunctionComponent = () => {
+export const SignUpStepper: FunctionComponent = () => {
   const { t } = useTranslation('yourloops')
   const { classes: { stepper } } = useStyles()
   const { logout } = useAuth()
-  const navigate = useNavigate()
+  const steps = ['select-account-type', 'consent', 'create-profile']
   const [activeStep, setActiveStep] = useState(0)
-  const steps = [
-    'select-account-type',
-    'consent',
-    'create-profile'
-  ]
 
   const handleNext = (): void => {
     setActiveStep((prevActiveStep) => prevActiveStep + 1)
@@ -78,23 +70,6 @@ const SignUpStepper: FunctionComponent = () => {
       setActiveStep((prevActiveStep) => prevActiveStep - 1)
     } else {
       logout()
-    }
-  }
-
-  const redirectToHome = (): void => {
-    navigate('/')
-  }
-
-  const getStepContent = (step: number): JSX.Element | string => {
-    switch (step) {
-      case 0:
-        return <SignUpAccountSelector handleBack={handleBack} handleNext={handleNext} />
-      case 1:
-        return <SignUpConsent handleBack={handleBack} handleNext={handleNext} />
-      case 2:
-        return <SignUpProfileForm handleBack={handleBack} handleNext={handleNext} />
-      default:
-        return t('signup-unknown-step')
     }
   }
 
@@ -123,39 +98,10 @@ const SignUpStepper: FunctionComponent = () => {
         ))}
       </Stepper>
 
-      {activeStep === steps.length
-        ? <Box paddingX={2} paddingTop={1} textAlign="left">
-          <Typography id="signup-steppers-ending-text-1" variant="h6" gutterBottom>
-            {t('account-creation-finalized')}
-          </Typography>
-          <Typography gutterBottom>
-            {t('account-created-info-1')}.
-          </Typography>
-          <Typography gutterBottom>
-            {t('account-created-info-2')}.
-          </Typography>
-          <Box
-            id="signup-consent-button-group"
-            display="flex"
-            justifyContent="center"
-            mx={0}
-            mt={4}
-          >
-            <Button
-              data-testid="validate-account-completion"
-              variant="contained"
-              color="primary"
-              disableElevation
-              onClick={redirectToHome}
-            >
-              {t('button-continue')}
-            </Button>
-          </Box>
-        </Box>
-        : getStepContent(activeStep)
-      }
+      {activeStep === 0 && <SignUpAccountSelector handleBack={handleBack} handleNext={handleNext} />}
+      {activeStep === 1 && <SignUpConsent handleBack={handleBack} handleNext={handleNext} />}
+      {activeStep === 2 && <SignUpProfileForm handleBack={handleBack} handleNext={handleNext} />}
+      {activeStep === 3 && <SignupCompleteMessage />}
     </React.Fragment>
   )
 }
-
-export default SignUpStepper